+++ charlatan/tests/fixtures/simple_models.py
@@ -1,0 +1,8 @@
+
+
+class Toaster(object):
+
+    def __init__(self, color, slots, content):
+        self.color = color
+        self.slots = slots
+        self.content = content