+++ charlatan/tests/test_relationships_without_models.py
@@ -3,15 +3,15 @@ from __future__ import absolute_import
 from charlatan import testcase, testing, FixturesManager
 
 
-class TestRelationshipsWithoutModels(testing.TestCase, testcase.FixturesMixin):
+class TestRelationshipsWithoutModels(testing.TestCase,
+                                     testcase.FixturesManagerMixin):
 
     fixtures = ('dict_with_nest', 'simple_dict', 'list_of_relationships',)
 
     def setUp(self):
-        fm = FixturesManager()
-        fm.load('./charlatan/tests/data/relationships_without_models.yaml')
-
-        self.use_fixtures_manager(fm)
+        self.fixtures_manager = FixturesManager()
+        self.fixtures_manager.load('./charlatan/tests/data/relationships_without_models.yaml')
+        self.init_fixtures()
 
     def test_dictionaries_nest(self):
         self.assertEqual(self.dict_with_nest['simple_dict'], self.simple_dict)